MP-312 & MP-313 update work exp modal
