+++ src/libs/shared/lib/constants/index.ts
@@ -1,0 +1,12 @@
+export const INDUSTRIES_OPTIONS: string[] = [
+    'Banking',
+    'Consumer goods',
+    'Energy',
+    'Entertainment',
+    'Healthcare',
+    'Pharma',
+    'Tech & technology services',
+    'Telecoms',
+    'Public sector',
+    'Travel & hospitality',
+]
